\name{combine.wsrf}
\alias{combine}
\alias{combine.wsrf}

\title{Combine Ensembles of Trees}

\description{
Combine two more more ensembles of trees into one.
}

\usage{
combine(...)
}

\arguments{
  \item{...}{two or more objects of class \code{randomForest}, to be
    combined into one.}
}

\value{
An object of class \code{wsrf}.
}

\seealso{\code{\link{subset}}}